refactor: #40 service 경로수정
